# docs/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "flaps HTTP API",
    "version": "0.1.0",
    "description": "Admin CRUD, SDK key management, SDK identity, OFREP evaluation and ruleset sync surface exposed by flaps-server. Generated by hand from build_router in crates/flaps-server/src/lib.rs (main@b4e15b1); see docs/spec/api-v1.md for the accompanying guide (authentication model, SSE contract, ETag semantics, error format)."
  },
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session bearer token minted by POST /login. Required by every /projects/** admin route (project, environment, flag, segment, flag-env-config and SDK key management)."
      },
      "sdkKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "SDK key bearer token. GET /sdk/whoami and the OFREP evaluation routes accept both server-kind and client-kind keys. The /sync/v1/* routes require a server-kind key; a client-kind key on those routes receives 403 Forbidden (see api-v1.md for the server-vs-client distinction, which this single security scheme does not model)."
      }
    },
    "parameters": {
      "ProjectParam": {
        "name": "project",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Project key (kebab-case)."
      },
...
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid authentication credentials.",
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded.",
        "headers": {
          "Retry-After": { "$ref": "#/components/headers/RetryAfterHeader" }
        },
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      },
      "InternalError": {
        "description": "Unexpected internal error (e.g. a persistence failure).",
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["username", "password"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "Opaque bearer token for subsequent Authorization: Bearer headers." },
          "expires_at": { "type": "string", "format": "date-time" }
        },
        "required": ["token", "expires_at"]
      },
...
    }
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "Verify credentials and mint a session token",
        "operationId": "postLogin",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Session token minted.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects": {
      "get": {
        "summary": "List all projects",
        "operationId": "listProjects",
        "security": [{ "adminSession": [] }],
        "responses": {
          "200": {
            "description": "All projects.",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{project}": {
      "get": {
        "summary": "Fetch a single project",
        "operationId": "getProject",
        "security": [{ "adminSession": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectParam" }],
        "responses": {
          "200": {
            "description": "The project.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETagHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "summary": "Upsert a project (idempotent)",
        "operationId": "putProject",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/IfMatchHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } }
        },
        "responses": {
          "200": {
            "description": "Project updated.",
            "headers": {
              "ETag": { "$ref": "#/components/headers/ETagHeader" },
              "X-Flaps-Warning": { "$ref": "#/components/headers/XFlapsWarningHeader" }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } }
          },
          "201": {
            "description": "Project created.",
            "headers": {
              "ETag": { "$ref": "#/components/headers/ETagHeader" },
              "X-Flaps-Warning": { "$ref": "#/components/headers/XFlapsWarningHeader" }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "412": { "$ref": "#/components/responses/PreconditionFailed" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete a project (cascades to all child entities)",
        "operationId": "deleteProject",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/IfMatchHeader" }
        ],
        "responses": {
          "204": { "description": "Project deleted." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "412": { "$ref": "#/components/responses/PreconditionFailed" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{project}/environments": {
      "get": {
        "summary": "List environments in a project",
        "operationId": "listEnvironments",
        "security": [{ "adminSession": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectParam" }],
        "responses": {
          "200": {
            "description": "All environments in the project.",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Environment" } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{project}/environments/{env}": {
      "get": {
        "summary": "Fetch a single environment",
        "operationId": "getEnvironment",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/EnvParam" }
        ],
        "responses": {
          "200": {
            "description": "The environment.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETagHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Environment" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "summary": "Upsert an environment",
        "operationId": "putEnvironment",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/EnvParam" },
          { "$ref": "#/components/parameters/IfMatchHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Environment" } } }
        },
        "responses": {
          "200": {
            "description": "Environment updated.",
            "headers": {
              "ETag": { "$ref": "#/components/headers/ETagHeader" },
              "X-Flaps-Warning": { "$ref": "#/components/headers/XFlapsWarningHeader" }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Environment" } } }
          },
          "201": {
            "description": "Environment created.",
            "headers": {
              "ETag": { "$ref": "#/components/headers/ETagHeader" },
              "X-Flaps-Warning": { "$ref": "#/components/headers/XFlapsWarningHeader" }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Environment" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "412": { "$ref": "#/components/responses/PreconditionFailed" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete an environment",
        "operationId": "deleteEnvironment",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/EnvParam" },
          { "$ref": "#/components/parameters/IfMatchHeader" }
        ],
        "responses": {
          "204": { "description": "Environment deleted." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "412": { "$ref": "#/components/responses/PreconditionFailed" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{project}/flags": {
      "get": {
        "summary": "List all flags in a project",
        "operationId": "listFlags",
        "security": [{ "adminSession": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectParam" }],
        "responses": {
          "200": {
            "description": "All flags in the project.",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Flag" } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{project}/flags/{flag}": {
      "get": {
        "summary": "Fetch a single flag",
        "operationId": "getFlag",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/FlagParam" }
        ],
        "responses": {
          "200": {
            "description": "The flag.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETagHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Flag" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "summary": "Upsert a flag",
        "operationId": "putFlag",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/FlagParam" },
          { "$ref": "#/components/parameters/IfMatchHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Flag" } } }
        },
        "responses": {
          "200": {
            "description": "Flag updated.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETagHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Flag" } } }
          },
          "201": {
            "description": "Flag created.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETagHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Flag" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "412": { "$ref": "#/components/responses/PreconditionFailed" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete a flag",
        "operationId": "deleteFlag",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/FlagParam" },
          { "$ref": "#/components/parameters/IfMatchHeader" }
        ],
        "responses": {
          "204": { "description": "Flag deleted." },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "412": { "$ref": "#/components/responses/PreconditionFailed" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{project}/segments": {
      "get": {
        "summary": "List all segments in a project",
        "operationId": "listSegments",
        "security": [{ "adminSession": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectParam" }],
        "responses": {
          "200": {
            "description": "All segments in the project.",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Segment" } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{project}/segments/{segment}": {
      "get": {
        "summary": "Fetch a single segment",
        "operationId": "getSegment",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/SegmentParam" }
        ],
        "responses": {
          "200": {
            "description": "The segment.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETagHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Segment" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "summary": "Upsert a segment",
        "operationId": "putSegment",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/SegmentParam" },
          { "$ref": "#/components/parameters/IfMatchHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Segment" } } }
        },
        "responses": {
          "200": {
            "description": "Segment updated.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETagHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Segment" } } }
          },
          "201": {
            "description": "Segment created.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETagHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Segment" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "412": { "$ref": "#/components/responses/PreconditionFailed" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete a segment (refused if any environment's flags still reference it)",
        "operationId": "deleteSegment",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/SegmentParam" },
          { "$ref": "#/components/parameters/IfMatchHeader" }
        ],
        "responses": {
          "204": { "description": "Segment deleted." },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "412": { "$ref": "#/components/responses/PreconditionFailed" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{project}/flags/{flag}/environments/{env}/config": {
      "get": {
        "summary": "Fetch a flag's configuration for one environment",
        "operationId": "getFlagEnvConfig",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/FlagParam" },
          { "$ref": "#/components/parameters/EnvParam" }
        ],
        "responses": {
          "200": {
            "description": "The configuration.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETagHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FlagEnvConfig" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "summary": "Upsert a flag's configuration for one environment",
        "operationId": "putFlagEnvConfig",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/FlagParam" },
          { "$ref": "#/components/parameters/EnvParam" },
          { "$ref": "#/components/parameters/IfMatchHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FlagEnvConfig" } } }
        },
        "responses": {
          "200": {
            "description": "Configuration updated.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETagHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FlagEnvConfig" } } }
          },
          "201": {
            "description": "Configuration created.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETagHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FlagEnvConfig" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "412": { "$ref": "#/components/responses/PreconditionFailed" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete a flag's configuration for one environment",
        "operationId": "deleteFlagEnvConfig",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/FlagParam" },
          { "$ref": "#/components/parameters/EnvParam" },
          { "$ref": "#/components/parameters/IfMatchHeader" }
        ],
        "responses": {
          "204": { "description": "Configuration deleted." },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "412": { "$ref": "#/components/responses/PreconditionFailed" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{project}/environments/{env}/keys": {
      "post": {
        "summary": "Issue a new SDK key scoped to a project/environment",
        "operationId": "postSdkKey",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/EnvParam" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateSdkKeyRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Key created. The raw secret is returned once and never exposed again.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateSdkKeyResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "get": {
        "summary": "List SDK keys for a project/environment scope",
        "operationId": "listSdkKeys",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/EnvParam" }
        ],
        "responses": {
          "200": {
            "description": "All SDK key records in scope (revoked and active alike).",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/SdkKeyRecord" } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/projects/{project}/environments/{env}/keys/{prefix}": {
      "delete": {
        "summary": "Revoke an SDK key by prefix",
        "description": "Soft-revokes the key. A no-op (still 204) if the key does not exist or is already revoked.",
        "operationId": "deleteSdkKey",
        "security": [{ "adminSession": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectParam" },
          { "$ref": "#/components/parameters/EnvParam" },
          { "$ref": "#/components/parameters/PrefixParam" }
        ],
        "responses": {
          "204": { "description": "Key revoked (or already absent)." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/InvalidBody" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/sdk/whoami": {
      "get": {
        "summary": "Identify the calling SDK key and check its rate limit",
        "operationId": "getWhoami",
        "security": [{ "sdkKey": [] }],
        "responses": {
          "200": {
            "description": "Key scope and kind.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WhoamiResponse" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/ofrep/v1/evaluate/flags": {
      "post": {
        "summary": "OFREP bulk flag evaluation",
        "description": "Evaluates every flag in the caller's (project, environment) scope. Reads solely from the in-memory compiled ruleset cache; supports If-None-Match/304 based on the ruleset content hash.",
        "operationId": "postEvaluateFlagsBulk",
        "security": [{ "sdkKey": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IfNoneMatchHeader" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EvaluationRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Bulk evaluation result. Also returned (with an empty flags array and metadata.version \"0\") when no ruleset is cached yet for the scope.",
            "headers": { "ETag": { "$ref": "#/components/headers/ETagHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BulkSuccessResponse" } } }
          },
          "304": { "description": "Not modified: the client's If-None-Match matches the current ruleset content hash. No body." },
          "400": {
            "description": "Request body is not valid JSON (bulkEvaluationFailure, INVALID_CONTEXT).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EvaluationFailureResponse" } } }
          },
          "401": {
            "description": "Missing or invalid SDK key.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EvaluationFailureResponse" } } }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "headers": { "Retry-After": { "$ref": "#/components/headers/RetryAfterHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EvaluationFailureResponse" } } }
          },
          "500": {
            "description": "Failed to parse the compiled ruleset (GENERAL).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EvaluationFailureResponse" } } }
          }
        }
      }
    },
    "/ofrep/v1/evaluate/flags/{key}": {
      "post": {
        "summary": "OFREP single flag evaluation",
        "description": "Evaluates one feature flag for the caller's (project, environment) scope. Reads solely from the in-memory compiled ruleset cache.",
        "operationId": "postEvaluateFlagSingle",
        "security": [{ "sdkKey": [] }],
        "parameters": [{ "$ref": "#/components/parameters/OfrepFlagKeyParam" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EvaluationRequest" } } }
        },
        "responses": {
          "200": {
            "description": "serverEvaluationSuccess.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SingleSuccessResponse" } } }
          },
          "400": {
            "description": "Request body is not valid JSON (evaluationFailure, INVALID_CONTEXT).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SingleErrorResponse" } } }
          },
          "401": {
            "description": "Missing or invalid SDK key.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EvaluationFailureResponse" } } }
          },
          "404": {
            "description": "flagNotFound: the flag key does not exist, or no ruleset is cached yet for the scope.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SingleErrorResponse" } } }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "headers": { "Retry-After": { "$ref": "#/components/headers/RetryAfterHeader" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EvaluationFailureResponse" } } }
          },
          "500": {
            "description": "Failed to parse the compiled ruleset, or an internal evaluation error (GENERAL).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SingleErrorResponse" } } }
          }
        }
      }
    },
...
  }
}